Issue #39: Restructured telegram bot code.
