+++ src/app/api/bot/message-echo.ts
@@ -1,0 +1,10 @@
+import { bot } from './core/botSinglton';
+
+// DEBUG: Mirror all the texts
+bot.on('message:text', async (ctx) => {
+  await ctx.reply(
+    // Send bold text
+    `*${ctx.message.text}*`,
+    { parse_mode: 'MarkdownV2' },
+  );
+});
